\name{ls_groups}
\alias{ls_groups}
\alias{ls_group_info}
\title{Inspect linked brushing groups}
\usage{
ls_groups()
ls_group_info(group)
}
\arguments{
\item{group}{A non-empty group name.}
}
\value{
\code{ls_groups()} returns a data frame. \code{ls_group_info()} returns a list.
}
\description{
Inspect active linked brushing groups and the plots associated with them.
}
