% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_regression.R
\name{ls_binary_regression_pairwise}
\alias{ls_binary_regression_pairwise}
\title{Pairwise comparisons for a categorical term in a binary regression}
\usage{
ls_binary_regression_pairwise(model, term, scale = NULL, adjust = "tukey")
}
\arguments{
\item{model}{Binary regression model.}

\item{term}{Simple categorical term.}

\item{scale}{For logit, \code{"odds_ratio"}, \code{"link"}, or \code{"probability"}; for
probit, \code{"probability"} or \code{"link"}.}

\item{adjust}{Multiplicity adjustment passed to \code{emmeans}.}
}
\value{
A data frame calculated by \code{emmeans} in R.
}
\description{
Pairwise comparisons for a categorical term in a binary regression
}
